move back changes for ScanUtils
