+++ src/main/java/io/github/bobocodebreskul/context/scan/utils/ScanUtilsImpl.java
@@ -1,7 +1,8 @@
-package io.github.bobocodebreskul.context;
+package io.github.bobocodebreskul.context.scan.utils;
 
 import static java.util.function.Predicate.not;
 
+import io.github.bobocodebreskul.context.scan.utils.ScanUtils;
 import java.lang.annotation.Annotation;
 import java.util.ArrayDeque;
 import java.util.Arrays;
